ai-backend/requirements.txt
Flask==2.3.3
torch==2.0.1
torchvision==0.15.2
opencv-python-headless==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
ultralytics==8.0.196